third_party/blink/web_tests/platform/generic/fast/events/page-visibility-iframe-unload-expected.txt

This test checks that the page visibility event is fired when frame is unloaded.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Loaded all frames.
PASS frameDocs[0].visibilityState is "hidden"
PASS frameDocs[1].visibilityState is "hidden"
PASS frameDocs[2].visibilityState is "hidden"
PASS successfullyParsed is true

TEST COMPLETE
